+++ source/sensor_types_info.sql
@@ -4,5 +4,5 @@ CREATE TABLE sensor_types_info
     id_type serial PRIMARY KEY,
     name varchar,
     description varchar,
-    error double
+    error float
 );
